refactor(cmd): extract run* subcommand handlers to internal/app; main is now thin dispatch
